app/requirements.txt
annotated-types==0.5.0
async-timeout==4.0.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
Flask==2.3.3
fonttools==4.42.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.8.0
midi2audio==0.1.1
mido==1.3.0
numba==0.57.1
numpy==1.24.4
packaging==23.1
Pillow==10.0.1
pretty-midi==0.2.10
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
spotipy==2.23.0
threadpoolctl==3.2.0
tslearn==0.6.2
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7